services/user-service/requirements.txt
# services/user-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0

# gRPC dependencies - UPDATED VERSIONS
grpcio==1.64.0
grpcio-tools==1.64.0
protobuf==5.27.1